Modifications 10 fevrier 2025

+++ database/migrations/2025_01_30_152445_create_etat_besoins_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->date('date_reception');
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->string('fichier'); // Ce champ va stocker le chemin du fichier PDF
-            $table->float('montant');
+            $table->decimal('montant', 15, 2);
             $table->timestamps();
         });
     }
